♻️ [Refactoring]: #455 Phase 2 basic カテゴリを tests/basic.rs に分離
空文字列 / 単純 ASCII / mid-buffer 系のテスト 5 件を
literal_string/tests/basic.rs に切り出し、tests.rs ハブに
`mod basic;` 宣言を追加。
